+++ arviz/stats/stats.py
@@ -693,6 +693,22 @@ def psislw(log_weights, reff=1.0):
         Smoothed log weights
     kss : array
         Pareto tail indices
+
+    References
+    ----------
+    * Vehtari et al. (2015) see https://arxiv.org/abs/1507.02646
+
+    Examples
+    --------
+    Get Pareto smoothed importance sampling (PSIS) log weights:
+
+    .. ipython::
+
+        In [1]: import arviz as az
+           ...: data = az.load_arviz_data("centered_eight")
+           ...: log_likelihood = data.sample_stats.log_likelihood.stack(sample=("chain", "draw"))
+           ...: az.psislw(-log_likelihood, reff=0.8)
+
     """
     if hasattr(log_weights, "sample"):
         n_samples = len(log_weights.sample)
@@ -874,6 +890,19 @@ def r2_score(y_true, y_pred):
     Pandas Series with the following indices:
     r2: Bayesian R²
     r2_std: standard deviation of the Bayesian R².
+
+    Examples
+    --------
+    Calculate R² for Bayesian regression models :
+
+    .. ipython::
+
+        In [1]: import arviz as az
+           ...: data = az.load_arviz_data('regression1d')
+           ...: y_true = data.observed_data["y"].values
+           ...: y_pred = data.posterior_predictive.stack(sample=("chain", "draw"))["y"].values.T
+           ...: az.r2_score(y_true, y_pred)
+
     """
     _numba_flag = Numba.numba_flag
     if y_pred.ndim == 1: